Update more affiliate page stuff
