+++ app/components/affiliate-page/affiliate-referred-user-item.hbs
@@ -1,4 +1,3 @@
-{{! @glint-nocheck: not typesafe yet }}
 <div
   class="flex items-center flex-wrap gap-y-2 justify-between group p-2 pr-3 hover:bg-gray-50 border rounded mb-2"
   data-test-affiliate-referred-user-item
@@ -11,7 +10,7 @@
   >
     <AvatarImage
       @user={{@affiliateReferral.customer}}
-      class="w-8 h-8 border border-gray-300 rounded shadow-sm group-hover:shadow mr-2 transition-shadow rounded-md"
+      class="w-8 h-8 border border-gray-300 rounded shadow-sm group-hover:shadow mr-2 transition-shadow"
     />
 
     <div>
