--- app/components/affiliate-page/affiliate-referred-user-item.js
@@ -1,3 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class AffiliateReferredUserItemComponent extends Component {}
