+++ app/components/affiliate-page/affiliate-referred-users-container.hbs
@@ -1,4 +1,3 @@
-{{! @glint-nocheck: not typesafe yet }}
 <PageSection @title="Referred Users" data-test-affiliate-referred-users-container ...attributes>
   {{#each this.paidReferrals as |affiliateReferral|}}
     <AffiliatePage::AffiliateReferredUserItem @affiliateReferral={{affiliateReferral}} />
