+++ app/templates/partner.hbs
@@ -18,7 +18,6 @@
         <AffiliatePage::AffiliateLinkStatsContainer @affiliateLink={{this.affiliateLink}} class="mt-4" />
         {{! @glint-expect-error: Not ts-ified yet }}
         <AffiliatePage::EarningsContainer class="mt-4" />
-        {{! @glint-expect-error: Not ts-ified yet }}
         <AffiliatePage::AffiliateReferredUsersContainer @affiliateLink={{this.affiliateLink}} class="mt-4" />
       </div>
